Updating alerts

+++ Frontend/src/features/exercises/Exercises.tsx
@@ -1,31 +1,26 @@
-import * as React from 'react';
+import React, { useState } from 'react';
+import { useAlert } from '../../contexts/AlertContext';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import Container from '@mui/material/Container';
 import ExerciseDialog from './ExerciseDialog';
 import Grid from '@mui/material/Grid';
-import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import Paper from '@mui/material/Paper';
-
 
 
 import AddIcon from '@mui/icons-material/Add';
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
 import type { Exercise } from '../../types';
-import SuccessAlert from '../shared/components/SuccessAlert';
-import ErrorAlert from '../shared/components/ErrorAlert';
-
+import { useConfirm } from '../../contexts/ConfirmContext';
 
 function Exercises() {
 
-  const [exercises, setExercises] = React.useState<Exercise[]>([]);
-  const [openExerciseDialog, setOpenExerciseDialog] = React.useState(false);
-  const [exerciseId, setExerciseId] = React.useState(0);
-
-  const [successMessage, setSuccessMessage] = React.useState("");
-  const [errorMessage, setErrorMessage] = React.useState("");
+  const [exercises, setExercises] = useState<Exercise[]>([]);
+  const [openExerciseDialog, setOpenExerciseDialog] = useState(false);
+  const [exerciseId, setExerciseId] = useState(0);
+  const { showAlert, showSuccess } = useAlert();
+  const { showConfirm } = useConfirm();
 
   function refreshExercises() {
     fetch("http://localhost:5103/exercises")
@@ -43,36 +38,41 @@ function Exercises() {
     setExerciseId(0);
     setOpenExerciseDialog(true);
   };
+
   const handleClose = () => {
     setOpenExerciseDialog(false);
     refreshExercises();
   };
 
 
-  // edit
   const handleEdit = (id: number) => {
     setExerciseId(id);
     setOpenExerciseDialog(true);
   }
 
 
-  // delete
   const handleDelete = async (id: number) => {
+
+    const confirmed = await showConfirm("Are you sure you want to delete this exercise?", "Confirm Delete");
+    if (!confirmed) {
+      return;
+    }
+    
     const url = `http://localhost:5103/exercises/${id}`;
     try {
       const response = await fetch(url, { method: 'DELETE' });
       if (!response.ok) {
-        setErrorMessage("Error deleting exercise!");
+        showAlert("Error deleting exercise!");
         console.error('Failed to delete exercise:', response.statusText);
         return;
       }
 
-      setSuccessMessage("Exercise deleted successfully!");
+      showSuccess("Exercise deleted successfully!");
       refreshExercises();
 
     } catch (error) {
       console.error('Error deleting exercise:', error);
-      setErrorMessage("Error deleting exercise!");
+      showAlert("Error deleting exercise!");
     }
   };
 
@@ -110,9 +110,6 @@ function Exercises() {
         ))}
       </Grid>
       <ExerciseDialog open={openExerciseDialog} onClose={handleClose} id={exerciseId} />
-
-      <SuccessAlert message={successMessage} />
-      <ErrorAlert message={errorMessage} />
     </>
   )
 }
